Import category position from Akeneo
