+++ install/Application/config/services_test.yaml
@@ -10,4 +10,5 @@ services:
         arguments:
             $categoryCodesToImport: []
             $categoryCodesToExclude: []
+            $useAkeneoPositions: false
         public: true
