+++ src/Form/Type/CategoriesConfigurationType.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Synolia\SyliusAkeneoPlugin\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\FormBuilderInterface;
 
@@ -25,6 +26,10 @@ final class CategoriesConfigurationType extends AbstractType
                 'label' => 'sylius.ui.admin.akeneo.categories.categories_to_exclude',
                 'required' => false,
                 'multiple' => true,
+            ])
+            ->add('use_akeneo_positions', CheckboxType::class, [
+                'label' => 'sylius.ui.admin.akeneo.categories.use_akeneo_positions',
+                'required' => false,
             ])
             ->add('submit', SubmitType::class, [
                 'attr' => ['class' => 'ui icon primary button'],
