+++ src/Processor/Category/CategoryProcessorChainInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Synolia\SyliusAkeneoPlugin\Processor\Category;
+
+use Sylius\Component\Core\Model\TaxonInterface;
+
+interface CategoryProcessorChainInterface
+{
+    public function chain(TaxonInterface $taxon, array $resource): void;
+}
