+++ src/Task/Category/CreateUpdateEntityTask.php
@@ -4,34 +4,22 @@ declare(strict_types=1);
 
 namespace Synolia\SyliusAkeneoPlugin\Task\Category;
 
-use Behat\Transliterator\Transliterator;
 use Doctrine\ORM\EntityManagerInterface;
-use Gedmo\Sortable\SortableListener;
 use Psr\Log\LoggerInterface;
 use Sylius\Component\Core\Model\TaxonInterface;
-use Sylius\Component\Resource\Factory\FactoryInterface;
-use Sylius\Component\Resource\Repository\RepositoryInterface;
 use Sylius\Component\Taxonomy\Factory\TaxonFactoryInterface;
-use Sylius\Component\Taxonomy\Model\TaxonTranslationInterface;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
-use Synolia\SyliusAkeneoPlugin\Builder\TaxonAttribute\TaxonAttributeValueBuilder;
-use Synolia\SyliusAkeneoPlugin\Component\TaxonAttribute\Model\TaxonAttributeSubjectInterface;
-use Synolia\SyliusAkeneoPlugin\Entity\TaxonAttribute;
-use Synolia\SyliusAkeneoPlugin\Entity\TaxonAttributeInterface;
-use Synolia\SyliusAkeneoPlugin\Entity\TaxonAttributeValueInterface;
 use Synolia\SyliusAkeneoPlugin\Event\Category\AfterProcessingTaxonEvent;
 use Synolia\SyliusAkeneoPlugin\Event\Category\BeforeProcessingTaxonEvent;
-use Synolia\SyliusAkeneoPlugin\Exceptions\UnsupportedAttributeTypeException;
 use Synolia\SyliusAkeneoPlugin\Logger\Messages;
+use Synolia\SyliusAkeneoPlugin\Manager\Doctrine\DoctrineSortableManager;
 use Synolia\SyliusAkeneoPlugin\Payload\Category\CategoryPayload;
 use Synolia\SyliusAkeneoPlugin\Payload\PipelinePayloadInterface;
+use Synolia\SyliusAkeneoPlugin\Processor\Category\CategoryProcessorChainInterface;
 use Synolia\SyliusAkeneoPlugin\Provider\Configuration\Api\CategoryConfigurationProviderInterface;
-use Synolia\SyliusAkeneoPlugin\Provider\SyliusAkeneoLocaleCodeProvider;
 use Synolia\SyliusAkeneoPlugin\Repository\TaxonRepository;
 use Synolia\SyliusAkeneoPlugin\Task\AkeneoTaskInterface;
-use Synolia\SyliusAkeneoPlugin\TypeMatcher\TaxonAttribute\TaxonAttributeTypeMatcher;
 use Throwable;
-use Webmozart\Assert\Assert;
 
 /**
  * @internal
@@ -44,25 +32,14 @@ final class CreateUpdateEntityTask implements AkeneoTaskInterface
 
     private string $type;
 
-    private array $taxonAttributes = [];
-
-    private array $taxonAttributeValues = [];
-
     public function __construct(
         private TaxonFactoryInterface $taxonFactory,
         private EntityManagerInterface $entityManager,
         private TaxonRepository $taxonRepository,
-        private RepositoryInterface $taxonTranslationRepository,
-        private FactoryInterface $taxonTranslationFactory,
         private LoggerInterface $logger,
         private EventDispatcherInterface $dispatcher,
-        private SyliusAkeneoLocaleCodeProvider $syliusAkeneoLocaleCodeProvider,
-        private RepositoryInterface $taxonAttributeRepository,
-        private RepositoryInterface $taxonAttributeValueRepository,
-        private FactoryInterface $taxonAttributeFactory,
-        private FactoryInterface $taxonAttributeValueFactory,
-        private TaxonAttributeTypeMatcher $taxonAttributeTypeMatcher,
-        private TaxonAttributeValueBuilder $taxonAttributeValueBuilder,
+        private DoctrineSortableManager $sortableManager,
+        private CategoryProcessorChainInterface $processorChain,
         private CategoryConfigurationProviderInterface $categoryConfigurationProvider,
     ) {
     }
@@ -76,6 +53,10 @@ final class CreateUpdateEntityTask implements AkeneoTaskInterface
         $this->type = $payload->getType();
         $this->logger->notice(Messages::createOrUpdate($this->type));
 
+        if (true === $this->categoryConfigurationProvider->get()->useAkeneoPositions()) {
+            $this->sortableManager->disableSortableEventListener();
+        }
+
         foreach ($payload->getResources() as $resource) {
             try {
                 $this->dispatcher->dispatch(new BeforeProcessingTaxonEvent($resource));
@@ -86,92 +67,9 @@ final class CreateUpdateEntityTask implements AkeneoTaskInterface
 
                 $taxon = $this->getOrCreateEntity($resource['code']);
 
-                $taxons[$resource['code']] = $taxon;
-
-                $this->assignParent($taxon, $taxons, $resource);
-
-                $this->logger->warning('SetPosition', [
-                    'taxon_code' => $taxon->getCode(),
-                    'position' => $resource['position'],
-                    'positiond' => $taxon->getPosition(),
-                ]);
-
-                foreach ($this->syliusAkeneoLocaleCodeProvider->getUsedLocalesOnBothPlatforms() as $syliusLocale) {
-                    $akeneoLocale = $this->syliusAkeneoLocaleCodeProvider->getAkeneoLocale($syliusLocale);
-
-                    $label = \sprintf('[%s]', $resource['code']);
-
-                    if (array_key_exists($akeneoLocale, $resource['labels']) && null !== $resource['labels'][$akeneoLocale]) {
-                        $label = $resource['labels'][$akeneoLocale];
-                    }
-
-                    $taxonTranslation = $this->taxonTranslationRepository->findOneBy([
-                        'translatable' => $taxon,
-                        'locale' => $syliusLocale,
-                    ]);
-
-                    if (!$taxonTranslation instanceof TaxonTranslationInterface) {
-                        /** @var TaxonTranslationInterface $taxonTranslation */
-                        $taxonTranslation = $this->taxonTranslationFactory->createNew();
-                        $taxonTranslation->setLocale($syliusLocale);
-                        $taxonTranslation->setTranslatable($taxon);
-                        $this->entityManager->persist($taxonTranslation);
-
-                        $this->logger->notice('Created TaxonTranslation', [
-                            'taxon_id' => $taxon->getId() ?? 'unknown',
-                            'taxon_code' => $taxon->getCode(),
-                            'locale' => $syliusLocale,
-                            'akeneo_locale' => $akeneoLocale,
-                        ]);
-                    }
-
-                    $taxonTranslation->setName($label);
-                    $slug = Transliterator::transliterate(
-                        str_replace(
-                            '\'',
-                            '-',
-                            sprintf(
-                                '%s-%s',
-                                $resource['code'],
-                                $label,
-                            ),
-                        ),
-                    );
-                    $taxonTranslation->setSlug($slug);
-
-                    $this->logger->notice('Update TaxonTranslation', [
-                        'taxon_id' => $taxon->getId() ?? 'unknown',
-                        'taxon_code' => $taxon->getCode(),
-                        'locale' => $syliusLocale,
-                        'akeneo_locale' => $akeneoLocale,
-                        'name' => $label,
-                        'slug' => $slug,
-                    ]);
-
-                    $this->handleAttributes($taxon, $resource);
-                }
+                $this->processorChain->chain($taxon, $resource);
 
                 $this->dispatcher->dispatch(new AfterProcessingTaxonEvent($resource, $taxon));
-
-                if ($this->categoryConfigurationProvider->get()->isUseAkeneoPositions()) {
-                    $taxon->setPosition($resource['position']);
-
-                    foreach ($this->entityManager->getEventManager()->getAllListeners() as $listenerTypes) {
-                        foreach ($listenerTypes as $listener) {
-                            if (get_class($listener) === SortableListener::class) {
-                                $this->entityManager->getEventManager()->removeEventListener([
-                                    'onFlush',
-                                    'loadClassMetadata',
-                                    'prePersist',
-                                    'postPersist',
-                                    'preUpdate',
-                                    'postRemove',
-                                    'postFlush',
-                                ], $listener);
-                            }
-                        }
-                    }
-                }
 
                 $this->entityManager->flush();
 
@@ -186,23 +84,11 @@ final class CreateUpdateEntityTask implements AkeneoTaskInterface
             }
         }
 
+        $this->sortableManager->enableSortableEventListener();
+
         $this->logger->notice(Messages::countCreateAndUpdate($this->type, $this->createCount, $this->updateCount));
 
         return $payload;
-    }
-
-    private function assignParent(TaxonInterface $taxon, array $taxons, array $resource): void
-    {
-        if (null === $resource['parent']) {
-            return;
-        }
-
-        $parent = $taxons[$resource['parent']] ?? null;
-
-        if (!$parent instanceof TaxonInterface) {
-            return;
-        }
-        $taxon->setParent($parent);
     }
 
     private function getOrCreateEntity(string $code): TaxonInterface
@@ -225,119 +111,5 @@ final class CreateUpdateEntityTask implements AkeneoTaskInterface
         $this->logger->info(Messages::hasBeenUpdated($this->type, (string) $taxon->getCode()));
 
         return $taxon;
-    }
-
-    private function handleAttributes(TaxonInterface $taxon, array $resource): void
-    {
-        if (!$taxon instanceof TaxonAttributeSubjectInterface) {
-            $this->logger->warning('Missing TaxonAttributeSubjectInterface implementation on Taxon', [
-                'taxon_code' => $taxon->getCode(),
-            ]);
-
-            return;
-        }
-
-        if (!array_key_exists('values', $resource)) {
-            return;
-        }
-
-        foreach ($resource['values'] as $attributeValue) {
-            try {
-                $taxonAttribute = $this->getTaxonAttributes(
-                    $attributeValue['attribute_code'],
-                    $attributeValue['type'],
-                );
-
-                $taxonAttributeValue = $this->getTaxonAttributeValues(
-                    $taxon,
-                    $taxonAttribute,
-                    $attributeValue['locale'],
-                );
-
-                $value = $this->taxonAttributeValueBuilder->build(
-                    $attributeValue['attribute_code'],
-                    $attributeValue['type'],
-                    $attributeValue['locale'],
-                    $attributeValue['channel'],
-                    $attributeValue['data'],
-                );
-
-                $taxonAttributeValue->setValue($value);
-            } catch (UnsupportedAttributeTypeException $e) {
-                $this->logger->warning($e->getMessage(), [
-                    'trace' => $e->getTrace(),
-                    'exception' => $e,
-                ]);
-            }
-        }
-    }
-
-    private function getTaxonAttributes(string $attributeCode, string $type): TaxonAttributeInterface
-    {
-        if (array_key_exists($attributeCode, $this->taxonAttributes)) {
-            return $this->taxonAttributes[$attributeCode];
-        }
-
-        $taxonAttribute = $this->taxonAttributeRepository->findOneBy(['code' => $attributeCode]);
-
-        if ($taxonAttribute instanceof TaxonAttribute) {
-            $this->taxonAttributes[$attributeCode] = $taxonAttribute;
-
-            return $taxonAttribute;
-        }
-
-        $matcher = $this->taxonAttributeTypeMatcher->match($type);
-
-        /** @var TaxonAttributeInterface $taxonAttribute */
-        $taxonAttribute = $this->taxonAttributeFactory->createNew();
-        $taxonAttribute->setCode($attributeCode);
-        $taxonAttribute->setType($type);
-        $taxonAttribute->setStorageType($matcher->getAttributeType()->getStorageType());
-        $taxonAttribute->setTranslatable(false);
-
-        $this->entityManager->persist($taxonAttribute);
-        $this->taxonAttributes[$attributeCode] = $taxonAttribute;
-
-        return $taxonAttribute;
-    }
-
-    private function getTaxonAttributeValues(
-        TaxonInterface $taxon,
-        TaxonAttributeInterface $taxonAttribute,
-        ?string $locale,
-    ): TaxonAttributeValueInterface {
-        Assert::string($taxon->getCode());
-        Assert::string($taxonAttribute->getCode());
-
-        if (
-            array_key_exists($taxon->getCode(), $this->taxonAttributeValues) &&
-            array_key_exists($taxonAttribute->getCode(), $this->taxonAttributeValues[$taxon->getCode()]) &&
-            array_key_exists($locale ?? 'unknown', $this->taxonAttributeValues[$taxon->getCode()][$taxonAttribute->getCode()])
-        ) {
-            return $this->taxonAttributeValues[$taxon->getCode()][$taxonAttribute->getCode()][$locale ?? 'unknown'];
-        }
-
-        $taxonAttributeValue = $this->taxonAttributeValueRepository->findOneBy([
-            'subject' => $taxon,
-            'attribute' => $taxonAttribute,
-            'localeCode' => $locale,
-        ]);
-
-        if ($taxonAttributeValue instanceof TaxonAttributeValueInterface) {
-            $this->taxonAttributeValues[$taxon->getCode()][$taxonAttribute->getCode()][$locale ?? 'unknown'] = $taxonAttributeValue;
-
-            return $taxonAttributeValue;
-        }
-
-        /** @var TaxonAttributeValueInterface $taxonAttributeValue */
-        $taxonAttributeValue = $this->taxonAttributeValueFactory->createNew();
-        $taxonAttributeValue->setAttribute($taxonAttribute);
-        $taxonAttributeValue->setTaxon($taxon);
-        $taxonAttributeValue->setLocaleCode($locale);
-        $this->entityManager->persist($taxonAttributeValue);
-
-        $this->taxonAttributeValues[$taxon->getCode()][$taxonAttribute->getCode()][$locale ?? 'unknown'] = $taxonAttributeValue;
-
-        return $taxonAttributeValue;
     }
 }
